Reworked helicopters
Helicopter throttle handling was changed. The previous function had the acceleration function anchored at 0.5 where 0.5 throttle was a hover. The new function reduces that to a very basic linear function where the engine power determines the slope and the function is anchored at the origin. The hover throttle is determined at runtime.